+++ Documentation/Changelog.md
@@ -50,6 +50,7 @@ For more details, please refer to:
 
 For more details, please refer to:
 
+* [MariaDB MaxScale 6.4.11 Release Notes](Release-Notes/MaxScale-6.4.11-Release-Notes.md)
 * [MariaDB MaxScale 6.4.10 Release Notes](Release-Notes/MaxScale-6.4.10-Release-Notes.md)
 * [MariaDB MaxScale 6.4.9 Release Notes](Release-Notes/MaxScale-6.4.9-Release-Notes.md)
 * [MariaDB MaxScale 6.4.8 Release Notes](Release-Notes/MaxScale-6.4.8-Release-Notes.md)